+++ rust/decentraland-godot-lib/src/scene_runner/components/mod.rs
@@ -11,8 +11,8 @@ pub mod mesh_collider;
 pub mod mesh_renderer;
 pub mod pointer_events;
 pub mod raycast;
-pub mod scene_ui;
 pub mod text_shape;
 pub mod transform_and_parent;
+pub mod ui;
 pub mod video_player;
 pub mod visibility;
